providers/gemini_provider.py
import google.generativeai as genai
import os
from typing import List, Dict, Any, Optional
from providers.base_provider import BaseLLMProvider

class GeminiProvider(BaseLLMProvider):
    """
    LLM Provider using Google Gemini.
    """
    def __init__(self, model_name: str = "gemini-2.5-flash", api_key: Optional[str] = None):
        self.api_key = api_key or os.environ.get("GOOGLE_API_KEY")
        if not self.api_key:
            raise ValueError("Google API Key not found. Please set GOOGLE_API_KEY environment variable.")
        genai.configure(api_key=self.api_key)
        self.model = genai.GenerativeModel(model_name)

    async def generate_response(
        self, 
        messages: List[Dict[str, str]], 
        tools: Optional[List[Dict[str, Any]]] = None
    ) -> Dict[str, Any]:
        """
        Generates a response using Gemini's API.
        Note: This is a simplified implementation for scaffolding.
        Gemini 1.5 supports function calling via tools.
        """
        # Convert messages to Gemini format
        history = []
        for msg in messages[:-1]:
            history.append({"role": "user" if msg["role"] == "user" else "model", "parts": [msg["content"]]})
        
        chat = self.model.start_chat(history=history)
        
        # Simplified handling: send the last message
        last_message = messages[-1]["content"]
        
        # In a real implementation with function calling, we would pass 'tools' here
        # For now, we'll just return text.
        response = await chat.send_message_async(last_message)
        
        return {
            "content": response.text,
            "tool_calls": [] # Placeholder for actual tool calls logic
        }
