requirements.txt
fastapi==0.135.1
uvicorn==0.42.0
google-generativeai==0.8.6
ollama==0.6.1
pydantic==2.12.5
python-dotenv==1.2.2
psutil==7.2.2
requests==2.32.5
dotenv==0.9.9
pypdf==6.8.0
beautifulsoup4==4.14.3
markdownify==1.2.0
httpx==0.28.1
playwright==1.58.0
crawl4ai==0.8.0
browser-use[video]
pandas==3.0.1
pyyaml==6.0.3
python-multipart==0.0.22
opencv-python==4.13.0.92
av==17.0.0
moviepy==2.2.1
selenium==3.141.0
Appium-Python-Client==1.3.0
urllib3==1.26.18